requirements.txt
requests==2.31.0
python-dotenv==1.0.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1
transformers==4.35.0 